Use only u64
Cosmos in general does not use u128 or bigint's unless it's specifically for
token balances. So we should use u64 for account numbers and sequence
numbers to match up with cosmos.

+++ src/client/get.rs
@@ -4,7 +4,7 @@ use crate::types::*;
 use deep_space::{address::Address, coin::Coin};
 
 impl Contact {
-    pub async fn get_latest_block_number(&self) -> Result<u128, JsonRpcError> {
+    pub async fn get_latest_block_number(&self) -> Result<u64, JsonRpcError> {
         let none: Option<bool> = None;
         let res: Result<LatestBlockEndpointResponse, JsonRpcError> = self
             .jsonrpc_client
